+++ src/navigation/action/preselectCharacterTwoColor.action.js
@@ -1,0 +1,8 @@
+export const PRESELECT_CHARACTER_TWO_COLOR = "preselectCharacterTwoColor";
+
+export default function preselectCharacterTwoColor(colorIndex) {
+  return {
+    type: PRESELECT_CHARACTER_TWO_COLOR,
+    colorIndex
+  };
+}