+++ src/navigation/navigation.reducer.js
@@ -2,12 +2,14 @@ import { PRESELECT_CHARACTER_ONE } from "./action/preselectCharacterOne.action";
 import { SELECT_CHARACTER_ONE } from "./action/selectCharacterOne.action";
 import { SELECT_CHARACTER_ONE_WITH_STYLE_AND_COLOR } from "./action/selectCharacterOneWithSyleAndColor.action";
 import { SELECT_CHARACTER_ONE_STYLE } from "./action/selectCharacterOneStyle.action";
+import { PRESELECT_CHARACTER_ONE_COLOR } from "./action/preselectCharacterOneColor.action";
 import { SELECT_CHARACTER_ONE_COLOR } from "./action/selectCharacterOneColor.action";
 import { UNSELECT_CHARACTER_ONE } from "./action/unselectCharacterOne.action";
 import { PRESELECT_CHARACTER_TWO } from "./action/preselectCharacterTwo.action";
 import { SELECT_CHARACTER_TWO } from "./action/selectCharacterTwo.action";
 import { SELECT_CHARACTER_TWO_WITH_STYLE_AND_COLOR } from "./action/selectCharacterTwoWithSyleAndColor.action";
 import { SELECT_CHARACTER_TWO_STYLE } from "./action/selectCharacterTwoStyle.action";
+import { PRESELECT_CHARACTER_TWO_COLOR } from "./action/preselectCharacterTwoColor.action";
 import { SELECT_CHARACTER_TWO_COLOR } from "./action/selectCharacterTwoColor.action";
 import { SELECT_CHARACTER_TWO_AI_LEVEL } from "./action/selectCharacterTwoAILevel.action";
 import { UNSELECT_CHARACTER_TWO } from "./action/unselectCharacterTwo.action";
@@ -18,12 +20,14 @@ import preselectCharacterOne from "./transition/preselectCharacterOne.transition
 import selectCharacterOne from "./transition/selectCharacterOne.transition";
 import selectCharacterOneWithStyleAndColor from "./transition/selectCharacterOneWithStyleAndColor.transition";
 import selectCharacterOneStyle from "./transition/selectCharacterOneStyle.transition";
+import preselectCharacterOneColor from "./transition/preselectCharacterOneColor.transition";
 import selectCharacterOneColor from "./transition/selectCharacterOneColor.transition";
 import unselectCharacterOne from "./transition/unselectCharacterOne.transition";
 import preselectCharacterTwo from "./transition/preselectCharacterTwo.transition";
 import selectCharacterTwo from "./transition/selectCharacterTwo.transition";
 import selectCharacterTwoWithStyleAndColor from "./transition/selectCharacterTwoWithStyleAndColor.transition";
 import selectCharacterTwoStyle from "./transition/selectCharacterTwoStyle.transition";
+import preselectCharacterTwoColor from "./transition/preselectCharacterTwoColor.transition";
 import selectCharacterTwoColor from "./transition/selectCharacterTwoColor.transition";
 import selectCharacterTwoAILevel from "./transition/selectCharacterTwoAILevel.transition";
 import unselectCharacterTwo from "./transition/unselectCharacterTwo.transition";
@@ -48,6 +52,9 @@ export default function navigationReducer(data, action) {
     case SELECT_CHARACTER_ONE_STYLE:
       return selectCharacterOneStyle(data, action);
 
+    case PRESELECT_CHARACTER_ONE_COLOR:
+      return preselectCharacterOneColor(data, action);
+
     case SELECT_CHARACTER_ONE_COLOR:
       return selectCharacterOneColor(data, action);
 
@@ -65,6 +72,9 @@ export default function navigationReducer(data, action) {
 
     case SELECT_CHARACTER_TWO_STYLE:
       return selectCharacterTwoStyle(data, action);
+
+    case PRESELECT_CHARACTER_TWO_COLOR:
+      return preselectCharacterTwoColor(data, action);
 
     case SELECT_CHARACTER_TWO_COLOR:
       return selectCharacterTwoColor(data, action);